requirements.txt
langchain
langchain-google-genai
langchain-chroma
chromadb
streamlit
python-dotenv